Fix the 'regex' converter so that parameters from the previous state can be reused
Docs for $state say: "If the state has an associated route, that route will be activated and the location with change it the address bar of the browser. It is also important that all parameters are defined for such route, however if the previous state defines any of those, they won't need to be redefined." As for now, the last part of the last sentence isn't true when the 'regex' converter is used.